// src/generated/model/soknadsStatusResponseStatus.ts
/**
 * Generated by orval v6.11.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type SoknadsStatusResponseStatus = typeof SoknadsStatusResponseStatus[keyof typeof SoknadsStatusResponseStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SoknadsStatusResponseStatus = {
    SENDT: "SENDT",
    MOTTATT: "MOTTATT",
    UNDER_BEHANDLING: "UNDER_BEHANDLING",
    FERDIGBEHANDLET: "FERDIGBEHANDLET",
    BEHANDLES_IKKE: "BEHANDLES_IKKE",
} as const;